Sort OpenAPI operation and path parameters
