+++ pkg/genericapiserver/openapi/openapi.go
@@ -183,8 +183,10 @@ func (o *openAPI) buildPaths() error {
 			for _, p := range inPathCommonParamsMap {
 				pathItem.Parameters = append(pathItem.Parameters, p)
 			}
+			sortParameters(pathItem.Parameters)
 			for _, route := range routes {
 				op, err := o.buildOperations(route, inPathCommonParamsMap)
+				sortParameters(op.Parameters)
 				if err != nil {
 					return err
 				}
@@ -285,28 +287,6 @@ func (o *openAPI) buildResponse(model interface{}, description string) (spec.Res
 			Schema:      schema,
 		},
 	}, nil
-}
-
-func groupRoutesByPath(routes []restful.Route) (ret map[string][]restful.Route) {
-	ret = make(map[string][]restful.Route)
-	for _, r := range routes {
-		route, exists := ret[r.Path]
-		if !exists {
-			route = make([]restful.Route, 0, 1)
-		}
-		ret[r.Path] = append(route, r)
-	}
-	return ret
-}
-
-func mapKeyFromParam(param *restful.Parameter) interface{} {
-	return struct {
-		Name string
-		Kind int
-	}{
-		Name: param.Data().Name,
-		Kind: param.Data().Kind,
-	}
 }
 
 func (o *openAPI) findCommonParameters(routes []restful.Route) (map[interface{}]spec.Parameter, error) {
@@ -411,55 +391,4 @@ func (o *openAPI) buildParameters(restParam []*restful.Parameter) (ret []spec.Pa
 		}
 	}
 	return ret, nil
-}
-
-// A simple trie implementation with Add an HasPrefix methods only.
-type trie struct {
-	children map[byte]*trie
-	wordTail bool
-}
-
-func createTrie(list []string) trie {
-	ret := trie{
-		children: make(map[byte]*trie),
-		wordTail: false,
-	}
-	for _, v := range list {
-		ret.Add(v)
-	}
-	return ret
-}
-
-func (t *trie) Add(v string) {
-	root := t
-	for _, b := range []byte(v) {
-		child, exists := root.children[b]
-		if !exists {
-			child = &trie{
-				children: make(map[byte]*trie),
-				wordTail: false,
-			}
-			root.children[b] = child
-		}
-		root = child
-	}
-	root.wordTail = true
-}
-
-func (t *trie) HasPrefix(v string) bool {
-	root := t
-	if root.wordTail {
-		return true
-	}
-	for _, b := range []byte(v) {
-		child, exists := root.children[b]
-		if !exists {
-			return false
-		}
-		if child.wordTail {
-			return true
-		}
-		root = child
-	}
-	return false
 }
